パッケージの追加と天気予報エンドポイントの実装
# 主要な変更内容
`Api.csproj`に新しいパッケージを追加し、`Program.cs`で天気予報エンドポイントの登録方法を変更しました。また、新たに`WeatherForecastEndpoints.cs`ファイルを追加し、天気予報データを生成してJSON形式で返す機能を実装しました。

## 変更したファイル
- `Api.csproj`: 新しいパッケージを追加
- `Program.cs`: 名前空間の追加、エンドポイント登録方法の変更
- `WeatherForecastEndpoints.cs`: 天気予報エンドポイントの実装

## 削除したファイル
- `WeatherForecast` クラスの定義
